Reworked maths core structure to be in separate files
