+++ mat_dp_core/maths_core/constraints.py
@@ -1,0 +1,63 @@
+from typing import Union
+
+import numpy as np
+from numpy import ndarray
+
+from .processes import Process, ProcessExpr
+
+
+def pack_constraint(constraint: Union[Process, ProcessExpr]) -> ndarray:
+    constraint *= 1  # Converts Process to ProcessExpr
+    proc_max_index = max((process.index for process in constraint)) + 1
+    array = np.zeros(proc_max_index)
+    for process in constraint:
+        array[process.index] = process.multiplier
+    return array
+
+
+class _Constraint:
+    name: str
+    array: ndarray
+    bound: float
+
+    def __init__(
+        self,
+        name: str,
+        weighted_processes: Union[Process, ProcessExpr],
+        bound: float,
+    ):
+        # TODO: investigate whether it would be nice to add a constraint by np.array
+        self.name = name
+        self.array = pack_constraint(weighted_processes)
+        self.bound = bound
+        self.weighted_processes = weighted_processes
+
+
+class EqConstraint(_Constraint):
+    """
+    Equality constraint
+    """
+
+    def __repr__(self) -> str:
+        return f"<EqConstraint: {self.name}\nEquation:{self.weighted_processes} == {self.bound}>"
+
+    def __format__(self, format_spec: str) -> str:
+        return f"{self.weighted_processes} == {self.bound}"
+
+
+class LeConstraint(_Constraint):
+    """
+    Less than or equal constraint
+    """
+
+    def __repr__(self) -> str:
+        return f"<LeConstraint: {self.name}\nEquation:{self.weighted_processes} == {self.bound}>"
+
+    def __format__(self, format_spec: str) -> str:
+        return f"{self.weighted_processes} == {self.bound}"
+
+
+def GeConstraint(
+    name: str, weighted_processes: Union[Process, ProcessExpr], bound: float
+):
+    return LeConstraint(name, -weighted_processes, -bound)
