+++ mat_dp_core/maths_core/measure.py
@@ -1,459 +1,26 @@
 import warnings
 from functools import reduce
-from itertools import starmap
-from typing import (
-    Any,
-    List,
-    MutableSequence,
-    Optional,
-    Sequence,
-    Tuple,
-    Union,
-    overload,
-)
+from typing import List, Optional, Sequence, Tuple, Union, overload
 
 import numpy as np
 from numpy import ndarray
 from scipy.optimize import linprog
 
-ResourceName = str
-Unit = str
-
-
-class Resource:
-    index: int
-    _parent: "Resources"
-
-    def __init__(self, index: int, parent: "Resources"):
-        self.index = index
-        self._parent = parent
-
-    @property
-    def name(self) -> str:
-        return self._parent._resources[self.index][0]
-
-    @property
-    def unit(self) -> str:
-        return self._parent._resources[self.index][1]
-
-    def __repr__(self):
-        return f"<Resource: {self.name}>"
-
-    def __eq__(self, other: Any) -> bool:
-        if not isinstance(other, Resource):
-            return False
-        else:
-            return self.index == other.index and self._parent == other._parent
-
-
-class Resources:
-    _resources: MutableSequence[Tuple[ResourceName, Unit]]
-
-    def __init__(self) -> None:
-        self._resources = []
-
-    def create(self, name: ResourceName, unit: Unit = "ea") -> Resource:
-        """
-        Create a resource
-        """
-        resource_out = self[len(self._resources)]
-        self._resources.append(
-            (
-                name,
-                unit,
-            )
-        )
-        return resource_out
-
-    def load(
-        self, resources: Sequence[Tuple[ResourceName, Unit]]
-    ) -> List[Resource]:
-        """
-        Load some additional resources in bulk
-        """
-        return list(starmap(self.create, resources))
-
-    def dump(self) -> Sequence[Tuple[ResourceName, Unit]]:
-        """
-        Dump resources in bulk
-        """
-        return self._resources
-
-    def __len__(self):
-        return len(self._resources)
-
-    def __getitem__(self, arg: Union[int, str]):
-        if isinstance(arg, int):
-            if arg > len(self._resources):
-                raise IndexError("list index out of range")
-            else:
-                return Resource(index=arg, parent=self)
-        else:
-            results = [
-                i for i, (name, _) in enumerate(self._resources) if name == arg
-            ]
-            if len(results) == 0:
-                raise KeyError(f"'{arg}'")
-            elif len(results) > 1:
-                raise KeyError(f"Name {arg} non unique: please use index")
-            else:
-                return Resource(index=results[0], parent=self)
-
-    def __iter__(self):
-        return map(self.__getitem__, range(len(self)))
-
-
-class ProcessExpr:
-    _processes: List["Process"]
-
-    def __init__(self, _processes: List["Process"]):
-        self._processes = _processes
-
-    def __add__(self, other: Union["ProcessExpr", "Process"]) -> "ProcessExpr":
-        if isinstance(other, ProcessExpr):
-            return ProcessExpr(self._processes + other._processes)
-        else:
-            return ProcessExpr(self._processes + [other])
-
-    def __mul__(self, other: float) -> "ProcessExpr":
-        if other == 1:
-            return self
-        for element in self._processes:
-            element.multiplier = element.multiplier * other
-        return self
-
-    def __rmul__(self, other: float) -> "ProcessExpr":
-        return self * other
-
-    def __neg__(self):
-        for element in self._processes:
-            element.multiplier = -element.multiplier
-        return self
-
-    def __sub__(self, other):
-        return self + -other
-
-    def __repr__(self) -> str:
-        return "<ProcessExpr {}>".format(
-            " + ".join(map(format, self._processes))
-        )
-
-    def __format__(self, format_spec: str) -> str:
-        return "{}".format(" + ".join(map(format, self._processes)))
-
-    def __getitem__(self, arg):
-        return self._processes[arg]
-
-    def __len__(self):
-        return len(self._processes)
-
-    def __iter__(self):
-        return map(self.__getitem__, range(len(self)))
-
-
-ProcessName = str
-
-
-class Process:
-    _parent: "Processes"
-    index: int
-    multiplier: float
-
-    def __init__(self, index: int, parent: "Processes"):
-        self._parent = parent
-        self.index = index
-        self.multiplier = 1
-
-    @property
-    def name(self) -> str:
-        return self._parent._processes[self.index][0]
-
-    @property
-    def array(self) -> ndarray:
-        return self._parent._processes[self.index][1]
-
-    def __mul__(self, other: float) -> ProcessExpr:
-        self.multiplier *= other
-        return ProcessExpr([self])
-
-    def __rmul__(self, other: float) -> ProcessExpr:
-        return self * other
-
-    def __add__(self, other: Union["Process", ProcessExpr]) -> ProcessExpr:
-        return self * 1 + other * 1
-
-    def __repr__(self) -> str:
-        return f"<Process: {self.name}" + (
-            ">" if self.multiplier == 1 else " * {self.multiplier}>"
-        )
-
-    def __format__(self, format_spec: str) -> str:
-        return f"{self.name}{'' if self.multiplier == 1 else f' * {self.multiplier}'}"
-
-    def __neg__(self):
-        return self * -1
-
-    def __sub__(self, other):
-        return self + -other
-
-    def __eq__(self, other: Any) -> bool:
-        if not isinstance(other, Process):
-            return False
-        else:
-            return self.index == other.index and self._parent == other._parent
-
-
-class Processes:
-    # Maps process names to resource demands
-    _processes: MutableSequence[Tuple[ProcessName, ndarray]]
-
-    def __init__(self) -> None:
-        self._processes = []
-
-    def create(
-        self, name: ProcessName, *resources: Tuple[Resource, float]
-    ) -> Process:
-        res_max_index = (
-            max((resource.index for (resource, _) in resources)) + 1
-        )
-        array = np.zeros(res_max_index)
-        for (resource, v) in resources:
-            i = resource.index
-            array[i] = v
-        process_inner = (name, array)
-        process_out = self[len(self._processes)]
-        self._processes.append(process_inner)
-        return process_out
-
-    def load(
-        self,
-        processes: Sequence[
-            Tuple[ProcessName, Sequence[Tuple[Resource, float]]]
-        ],
-    ) -> List[Process]:
-        """
-        Load some additional processes in bulk
-        """
-        return list(
-            starmap(
-                self.create,
-                [
-                    [process_name, *resources]
-                    for process_name, resources in processes
-                ],
-            )
-        )
-
-    def dump(self) -> Sequence[Tuple[ProcessName, ndarray]]:
-        """
-        Dump processes in bulk
-        """
-        return self._processes
-
-    def __len__(self):
-        return len(self._processes)
-
-    def __getitem__(self, arg: Union[int, str]):
-        if isinstance(arg, int):
-            if arg > len(self._processes):
-                raise IndexError("list index out of range")
-            else:
-                return Process(index=arg, parent=self)
-        else:
-            results = [
-                i for i, (name, _) in enumerate(self._processes) if name == arg
-            ]
-            if len(results) == 0:
-                raise KeyError(f"'{arg}'")
-            elif len(results) > 1:
-                raise KeyError(f"Name {arg} non unique: please use index")
-            else:
-                return Process(index=results[0], parent=self)
-
-    def __iter__(self):
-        return map(self.__getitem__, range(len(self)))
-
-
-def pack_constraint(constraint: Union[Process, ProcessExpr]) -> ndarray:
-    constraint *= 1  # Converts Process to ProcessExpr
-    proc_max_index = max((process.index for process in constraint)) + 1
-    array = np.zeros(proc_max_index)
-    for process in constraint:
-        array[process.index] = process.multiplier
-    return array
-
-
-class _Constraint:
-    name: str
-    array: ndarray
-    bound: float
-
-    def __init__(
-        self,
-        name: str,
-        weighted_processes: Union[Process, ProcessExpr],
-        bound: float,
-    ):
-        # TODO: investigate whether it would be nice to add a constraint by np.array
-        self.name = name
-        self.array = pack_constraint(weighted_processes)
-        self.bound = bound
-        self.weighted_processes = weighted_processes
-
-
-class EqConstraint(_Constraint):
-    """
-    Equality constraint
-    """
-
-    def __repr__(self) -> str:
-        return f"<EqConstraint: {self.name}\nEquation:{self.weighted_processes} == {self.bound}>"
-
-    def __format__(self, format_spec: str) -> str:
-        return f"{self.weighted_processes} == {self.bound}"
-
-
-class LeConstraint(_Constraint):
-    """
-    Less than or equal constraint
-    """
-
-    def __repr__(self) -> str:
-        return f"<LeConstraint: {self.name}\nEquation:{self.weighted_processes} == {self.bound}>"
-
-    def __format__(self, format_spec: str) -> str:
-        return f"{self.weighted_processes} == {self.bound}"
-
-
-def GeConstraint(
-    name: str, weighted_processes: Union[Process, ProcessExpr], bound: float
-):
-    return LeConstraint(name, -weighted_processes, -bound)
-
-
-class IterationLimitReached(Exception):
-    def __init__(self, n_iters):
-        self.n_iters = n_iters
-        super().__init__(f"Iteration limit reached with {n_iters} iterations")
-
-
-class Overconstrained(Exception):
-    def __init__(
-        self,
-        rec_constraints: Sequence[
-            Tuple[Resource, float, List[Process], List[Process]]
-        ],
-        eq_constraints: Sequence[Tuple[EqConstraint, float]],
-        le_constraints: Sequence[Tuple[LeConstraint, float]],
-    ):
-        def constraints_to_rec_string(
-            constraints: Sequence[
-                Tuple[Resource, float, List[Process], List[Process]]
-            ]
-        ) -> str:
-            rec_strings = []
-            for (constraint, val, producers, consumers) in constraints:
-                producer_names = ", ".join([p.name for p in producers])
-                consumer_names = ", ".join([p.name for p in consumers])
-                if val < 0:
-                    format_str = f"Increase runs of {producer_names} or decrease runs of {consumer_names}"
-                elif val > 0:
-                    format_str = f"Increase runs of {consumer_names} or decrease runs of {producer_names}"
-                else:
-                    format_str = None
-                if format_str is not None:
-                    full_string = (
-                        f"{constraint} => {val}: "
-                        + format_str.format(constraint.name)
-                    )
-                    rec_strings.append(full_string)
-            return "\n".join(rec_strings)
-
-        self.rec_constraints = rec_constraints
-        self.eq_constraints = eq_constraints
-        self.le_constraints = le_constraints
-        # TODO Fix residual val being non-zero
-        message_list = ["Overconstrained problem:"]
-        if len(rec_constraints) > 0:
-            message_list += [
-                "resources:",
-                constraints_to_rec_string(rec_constraints),
-            ]
-        if len(eq_constraints) > 0:
-            message_list += [
-                "eq_constraints:",
-                "\n".join([f"{c} => {val}" for (c, val) in eq_constraints]),
-            ]
-        if len(le_constraints) > 0:
-            message_list += "le_constraints:", "\n".join(
-                [f"{c} => {val}" for (c, val) in le_constraints]
-            )
-
-        super().__init__("\n".join(message_list))
-
-
-class UnboundedSolution(Exception):
-    def __init__(self, process_sols: List[Tuple[Process, float]]) -> None:
-        message_list = ["Solution unbounded - final solution was:"]
-        for p, sol in process_sols:
-            if sol > 10 ** 7 or sol < 10 ** -7:
-                comment = "(probably unbounded)"
-            else:
-                comment = ""
-            message_list.append(f"{p}: {sol} {comment}")
-
-        super().__init__("\n".join(message_list))
-
-
-class NumericalDifficulties(Exception):
-    pass
-
-
-class InconsistentOrderOfMagnitude(Exception):
-    def __init__(
-        self,
-        resources: List[Tuple[Resource, List[Tuple[Process, float]], float]],
-        eq_constraints: List[
-            Tuple[EqConstraint, List[Tuple[Process, float]], float]
-        ],
-        le_constraints: List[
-            Tuple[LeConstraint, List[Tuple[Process, float]], float]
-        ],
-    ) -> None:
-        message_list = [
-            "\nAll resources and constraints must be of a \nconsistent order of magnitude. If you wish to allow this \nbehaviour use 'allow_inconsistent_order_of_mag'.\n"
-        ]
-        if len(resources) > 0:
-            message_list.append("Resource inconsistencies")
-            for resource, process_list, order_range in resources:
-                message_list.append(
-                    f"{resource}: Order of mag range: {order_range}"
-                )
-                for process, process_demand in process_list:
-                    message_list.append(f"{process}: {process_demand}")
-                message_list.append("\n")
-            message_list.append("\n")
-
-        if len(eq_constraints) > 0:
-            message_list.append("Eq Constraint inconsistencies")
-            for eq_constraint, process_list, order_range in resources:
-                message_list.append(
-                    f"{eq_constraint}: Order of mag range - {order_range}"
-                )
-                for process, process_demand in process_list:
-                    message_list.append(f"{process}: {process_demand}")
-                message_list.append("\n")
-
-        if len(le_constraints) > 0:
-            message_list.append("Le Constraint inconsistencies")
-            for le_constraint, process_list, order_range in resources:
-                message_list.append(
-                    f"{le_constraint}: Order of mag range - {order_range}"
-                )
-                for process, process_demand in process_list:
-                    message_list.append(f"{process}: {process_demand}")
-                message_list.append("\n")
-        super().__init__("\n".join(message_list))
+from .constraints import (
+    EqConstraint,
+    LeConstraint,
+    _Constraint,
+    pack_constraint,
+)
+from .exceptions import (
+    InconsistentOrderOfMagnitude,
+    IterationLimitReached,
+    NumericalDifficulties,
+    Overconstrained,
+    UnboundedSolution,
+)
+from .processes import Process, Processes, ProcessExpr
+from .resources import Resource, Resources
 
 
 class Measure:
